Add API wrapper method for searching accounts
This commit adds the first of many API wrapper methods and introduces
ExVCR for caching API responses. We've added a git pre-commit hook to
ensure that our application_id is not committed in these fixture files.

The `pre-commit.sample` file needs to be moved to
`.git/hooks/pre-commit` to have any effect.